Renamed to a service, moved into folder and added a Model folder for the future database and logging layer
