+++ Stores/NavigationStore.cs
@@ -1,4 +1,4 @@
-﻿using Mayuri.Models;
+﻿using Mayuri.Services;
 using Mayuri.ViewModels;
 using System;
 using System.Collections.Generic;
