+++ ViewModels/MainViewModel.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using Mayuri.Models;
+using Mayuri.Services;
 using Mayuri.Stores;
 
 namespace Mayuri.ViewModels
