+++ ViewModels/MenuViewModel.cs
@@ -1,5 +1,5 @@
 ﻿using Mayuri.Commands;
-using Mayuri.Models;
+using Mayuri.Services;
 using Mayuri.Stores;
 using Mayuri.Views;
 using System;
